+++ src/SkyConnect/src/SkyConnectImpl.h
@@ -48,8 +48,9 @@ public:
 protected:
     virtual void onStartDataSample() override;
     virtual void onStopDataSample() override;
-    virtual void onStartReplay(bool fromStart) override;
+    virtual void onStartReplay(qint64 currentTimestamp) override;
     virtual void onStopReplay() override;
+    virtual void onSeek(qint64 currentTimestamp) override;
     virtual void onRecordingPaused(bool paused) override;
     virtual void onReplayPaused() override;
     virtual void onRecordSampleRateChanged(SampleRate::SampleRate sampleRate) override;
@@ -68,6 +69,7 @@ private:
     bool reconnectWithSim();
     bool close();
     void setupRequestData();
+    void setupInitialPosition();
     void setSimulationFrozen(bool enable);
     bool isSimulationFrozen() const;
     bool sendAircraftData();